breadcrumbs_php
 A function that produces a category breadcrumb. This function should
use parameters for the required category id and the separator.
For example. If I supplied 8 as the category id and ‘>’ as the separator I would expect the result to
be:
Clothing > Men > T-shirts
If I supplied 5 as the category id and ‘/’ as the separator I would expect the result to be:
Clothing / Men
For the purposes of this exercise assume you have a function called getCategory($id) available to
supply you with an array of the category values when the category id is passed in.